backend/requirements.txt
# Django core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# Authentication & Security
djangorestframework-simplejwt==5.3.1
argon2-cffi==23.1.0
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9

# Async / WebSocket
daphne==4.1.0
channels==4.1.0
channels-redis==4.2.0

# Task queue
celery==5.3.6
redis==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1

# File processing & validation
python-magic==0.4.27
pillow==10.3.0

# PDF generation
weasyprint==61.2
reportlab==4.1.0

# Templating
Jinja2==3.1.4

# Charts & data
matplotlib==3.8.4
plotly==5.21.0
pandas==2.2.1
kaleido==0.2.1  # Plotly static image export

# Email
django-anymail==10.3

# Utils
python-decouple==3.8
django-extensions==3.2.3
django-storages==1.14.2

# Testing
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Dev tools
black==24.3.0
isort==5.13.2
flake8==7.0.0
